Added methods in toolbar.

+++ src/org/hsm/view/MainFrame.java
@@ -87,7 +87,7 @@ public class MainFrame implements View {
     }
 
     @Override
-    public void insertPlant(final int id, final String name, final int cost, final int ph, final double brightness,
+    public void insertPlant(final int id, final String name, final double cost, final int ph, final double brightness,
                 final double conductivity, final double temperature) {
         this.tab.getPlantsTab().insertRow(id, name, cost, ph, brightness, conductivity, temperature);
     }
